Fix warnings.
